feat(actions,vendo): any REST API with a spec becomes agent tools
openApiConnector({ spec, baseUrl, headers, name }) turns an OpenAPI
document into guarded tools. It is reuse, not new machinery: the same
extractor `vendo sync` runs over a spec file, and the same HTTP dispatch
a host tool executes through.

Two factorings made that sharing possible:

- extractOpenApi's document half moved to the pure src/openapi-document.ts
  (the binding-identity.ts precedent). sync/openapi.ts keeps node:fs and
  the spec-file entry points; the connector is handed the document in
  memory. Keeping it in sync/ would have dragged sync/common.ts and its
  TypeScript compiler into the runtime entry, which the portability gate
  forbids outright (FORBIDDEN_INPUTS: packages/actions/dist/sync/).
  Route naming and extractedRisk moved alongside it into
  binding-identity.ts, re-exported from sync/common.ts.
- registry.ts's HTTP leg — argument binding, path substitution, the tRPC
  envelope, the fetch — is runtime/http-dispatch.ts now, used by the
  registry and the connector both. The JSON accept/content-type envelope
  moved inside fetchHostTool, so neither caller sets it.

McpAuthContext and McpHeadersResolver stay working as deprecated aliases
of ConnectorAuthContext / ConnectorHeadersResolver. Both connectors are
re-exported from @vendoai/vendo/server (and so from vendoai/server), and
docs-site/capabilities/connectors.mdx documents them — mcpConnector's
first page.

The test stands up a live HTTP fixture, points the connector at a spec
describing it, and executes through createActions: no stub on either
side. It proves the round trip, that baseUrl beats servers[0], that a
headers resolver sees the principal and grant, and that risk tracks the
method.
