+++ packages/actions/src/index.ts
@@ -13,6 +13,11 @@ export { composioConnector } from "./connectors/composio.js";
 export { composioToolRisk } from "./connectors/composio-risk.js";
 export { normalizeToolName } from "./connectors/names.js";
 export { mcpConnector, type McpAuthContext, type McpHeadersResolver } from "./connectors/mcp.js";
+export {
+  openApiConnector,
+  type ConnectorAuthContext,
+  type ConnectorHeadersResolver,
+} from "./connectors/openapi.js";
 export { createActions, type ActionsRegistry, type ActionsRunContext, type ServerActionHandler } from "./runtime/registry.js";
 export { createConnectGate, type ConnectGate, type ConnectGateOptions } from "./runtime/connect-gate.js";
 export { type ToolSearchMatch, type ToolSearchOptions } from "./runtime/search.js";
